chore(setup): move plugin initialization logic into setup/ directory
